+++ src/aux-server/aux-web/shared/scene/decorators/LabelDecorator.ts
@@ -7,6 +7,7 @@ import {
     calculateFormattedFileValue,
     calculateNumericalTagValue,
     getFileLabelAnchor,
+    FileLabelAnchor,
 } from '@casual-simulation/aux-common';
 import { Text3D } from '../Text3D';
 import { Color, Vector3, Box3, PerspectiveCamera } from 'three';
@@ -41,9 +42,37 @@ export class LabelDecorator extends AuxFile3DDecorator
     fileUpdated(calc: FileCalculationContext): void {
         let label = this.file3D.file.tags['aux.label'];
 
+        const anchor: FileLabelAnchor = calculateFileValue(
+            calc,
+            this.file3D.file,
+            'aux.label.anchor'
+        );
+
+        let botWidth = calculateNumericalTagValue(
+            calc,
+            this.file3D.file,
+            'aux.scale.x',
+            1
+        );
+
+        if (anchor === 'left' || anchor === 'right') {
+            botWidth = calculateNumericalTagValue(
+                calc,
+                this.file3D.file,
+                'aux.scale.y',
+                1
+            );
+        }
+
+        if (this.text3D) {
+            if (botWidth != this.text3D.currentWidth) {
+                this.disposeText3D();
+            }
+        }
+
         if (label) {
             if (!this.text3D) {
-                this.text3D = new Text3D();
+                this.text3D = new Text3D(botWidth * 100);
                 // Parent the labels directly to the file.
                 // Labels do all kinds of weird stuff with their transforms, so this makes it easier to let them do that
                 // without worrying about what the AuxFile3D scale is etc.