+++ src/aux-server/aux-web/shared/vue-components/Loading/Loading.ts
@@ -1,7 +1,7 @@
 import Component from 'vue-class-component';
 import Vue from 'vue';
 import { LoadingProgress } from '@casual-simulation/aux-common/LoadingProgress';
-import { appManager } from '../../../shared/AppManager';
+import { appManager, AppType } from '../../../shared/AppManager';
 import { hasValue } from '@casual-simulation/aux-common';
 import { Prop } from 'vue-property-decorator';
 import Tagline from '../Tagline/Tagline';
@@ -19,6 +19,10 @@ export default class Loading extends Vue {
 
     get hasError(): boolean {
         return hasValue(this.error);
+    }
+
+    get isPlayer(): boolean {
+        return appManager.appType === AppType.Player;
     }
 
     onErrorDismiss() {