+++ src/aux-server/server/webpack.prod.js
@@ -1,0 +1,14 @@
+const path = require('path');
+const merge = require('webpack-merge');
+const webpack = require('webpack');
+const TerserPlugin = require('terser-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const common = require('./webpack.common.js');
+
+module.exports = merge.smart(common, {
+    mode: 'production',
+    optimization: {
+        minimize: false,
+    },
+});